refact : refactoring signup page
