+++ package.json
@@ -55,5 +55,5 @@
 		"eslint-plugin-prettier": "^4.2.1",
 		"prettier": "^2.7.1"
 	},
-	"proxy": "http://ec2-3-34-75-204.ap-northeast-2.compute.amazonaws.com:8080"
+	"proxy": "http://localhost:8080/"
 }
