+++ src/components/Home/RefreshCircle/RefreshCircle.js
@@ -1,10 +1,12 @@
 import * as S from './RefreshCircle.style';
 import React from 'react';
 
-const RefreshCircle = () => {
+const RefreshCircle = ({ onClick }) => {
 	return (
 		<S.RefreshCircle>
-			<S.RefreshCircleInner className="material-icons">autorenew</S.RefreshCircleInner>
+			<S.RefreshCircleInner className="material-icons" onClick={onClick}>
+				autorenew
+			</S.RefreshCircleInner>
 		</S.RefreshCircle>
 	);
 };
