+++ src/components/Signup/PartnerMoreInfo/PartnerMoreInfo.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+
 import { NavigateButton } from '../../Button/Button';
 import Wrapper from '../../Wrapper/Wrapper';
 import * as S from './PartnerMoreInfo.style';
