--- src/components/Signup/PartnerMoreInfo/index.js
@@ -1,3 +1,0 @@
-import PartnerMoreInfo from './PartnerMoreInfo';
-
-export default PartnerMoreInfo;
