--- src/components/Signup/ProfilePhoto/index.js
@@ -1,2 +1,0 @@
-import ProfilePhoto from './ProfilePhoto';
-export default ProfilePhoto;
