+++ src/pages/SignUp/PartnerMoreInfo/PartnerMoreInfoPage.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import PartnerMoreInfo from '../../../components/Signup/PartnerMoreInfo';
+
+import PartnerMoreInfo from '../../../components/Signup/PartnerMoreInfo/PartnerMoreInfo';
 
 const PartnerMoreInfoPage = () => {
 	return <PartnerMoreInfo />;
