--- src/pages/SignUp/PartnerMoreInfo/index.js
@@ -1,3 +1,0 @@
-import PartnerMoreInfoPage from './PartnerMoreInfoPage';
-
-export default PartnerMoreInfoPage;
