+++ src/pages/SignUp/ProfilePhoto/ProfilePhotoPage.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import ProfilePhoto from '../../../components/Signup/ProfilePhoto/ProfilePhoto';
 
 const ProfilePhotoPage = () => {
