--- src/pages/SignUp/ProfilePhoto/index.js
@@ -1,3 +1,0 @@
-import ProfilePhotoPage from './ProfilePhotoPage';
-
-export default ProfilePhotoPage;
