+++ src/pages/Wait.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import Wrapper from '../components/Wrapper';
-import { WrapperInner } from '../components/Wrapper/Wrapper.style';
 import styled from 'styled-components';
-import { Red } from '../components/Signup/PartnerMoreInfo/PartnerMoreInfo.style';
+
 import { BigButton } from '../components/Button/BigButton/BigButton.style';
 import SideBar from '../components/SideBar/SideBar';
+import { Red } from '../components/Signup/PartnerMoreInfo/PartnerMoreInfo.style';
+import Wrapper from '../components/Wrapper';
+import { WrapperInner } from '../components/Wrapper/Wrapper.style';
 
 const Title = styled.section`
 	font-size: 2.6rem;
